requirements.txt
torch
torchvision
opencv-python
matplotlib
python-chess
IPython
numpy
